+++ GitEasy.psm1
@@ -5,9 +5,12 @@ Get-ChildItem -LiteralPath $privatePath -Filter '*.ps1' -File | Sort-Object Name
 Get-ChildItem -LiteralPath $publicPath -Filter '*.ps1' -File | Sort-Object Name | ForEach-Object { . $_.FullName }
 Export-ModuleMember -Function @(
     'Clear-Junk',
+    'Disable-GitEasy',
+    'Enable-GitEasy',
     'Find-CodeChange',
     'Get-Updates',
     'Get-VaultStatus',
+    'New-BugReport',
     'New-Release',
     'New-WorkBranch',
     'Reset-Login',